Add support for using tap adapters on unix. This allows iked to operate in virtual adapter mode on the unix platform.
Add support for route management on BSD platforms. This includes support for looking up, adding and removing routes. At the moment all routes are treated as network routes but this could easily be corrected in the future. The route increment and decrement functions have not been completed as Im still not sure how to handle this correctly.