Add workaround in iked for gateways that neglect to include an XAuth type attribute. Reported by Hiren Joshi.
Improve the shutdown procedure for iked, ipsecd and dtpd. Some issues with the code were preventing it from stopping correctly when acting as a service. This needs to be looked at closer once all the IPC work has settled.

Correct access control assignments for the generic IPC layer. This issue was preventing Vista clients from connecting to named pipe instances.