Modify iked to work around an issue in ipsec tools version 0.7. The unity split network data sent during modecfg can contain bogus port and protocol information. Although I have committed a fix for this in both the 0.7 and head ipsec tools branches, the code will need to settle for some time before the work around can be removed.
Modify iked to check return values properly inside the IPC server connection thread loop. This was causing the program to hang during shutdown when running as a service.

Modify libith to correct a bug in the windows version of the IPC server class. When creating a named pipe with explicit access, you must specify FILE_CREATE_PIPE_INSTANCE for an SID that is appropriate for the account that owns your process. Otherwise, after creating the initial pipe instance and assigning the access control, your process will loose its ability to create more than one pipe instance.