Modify a name service configuration structure data member used by iked. This is to accurately denote it is related to DNS.
Modify iked to improve logging while processing an inbound or outbound packet. We now include the ISAKMP SA cookies values.

Correct an issue in iked where we were responding to a new quick mode negotiation using a different ISAKMP SA than the initiator. While technically this is a perfectly legal thing to do, Cisco gateways ( and possibly others ) treat this as an error condition. The problem could be easily reproduced by configuring the phase2 lifetime values to be equal to the phase1 lifetime. We now store the ISAKMP cookie values used to initiate a quick mode exchange in the phase2 handle and use them to lookup the phase1 SA handle before responding. If the SA is not found, potentially due to lifetime expiration, we remove the phase2 handle and wait for a new negotiation. In my testing, this substantially improves the connection stability when no traffic is being sent by the client.