Revert the change that re-transmits a configuration packet after the configuration has become mature. This can cause problems when communicating with a Cisco device if the virtual network adapter is taking an unusually long time to initialize. A ping-pong packet war will commence which quickly leads to premature tunnel termination. A more appropriate long term fix will be committed after pre-requisite work is completed.
Introduce a new policy option that forces a single phase2 SA to be negotiated after a policy is created. Use this option to ensure phase2 negotiation occurs immediately after a connection has been established with a Cisco gateway. These devices will disconnect a client shortly after initial contact if an IPsec SA is not established.