Open the pfdev before we setuid/setgid. Otherwise, we wont be able to open the device.
Cache the syslog log service name or we won't have it for log rotation.

Display log output on stdout so we see errors at startup.

Output an error and exit if a configuraiton file cannot be loaded.

Fix a problem with setuid/setgid.

Add an install target to the makefile.